Refactor: Split out ast fragments from cpp.py
